+++ tests/get-tests.php
@@ -16,8 +16,10 @@ test( 'get', function () {
 
 test( 'get: timeout', function () {
 	$http = new \Amulet\HTTP\Request();
-	$http->timeout = 1;
-	$response = $http->get( url: 'http://127.0.0.1:7878/?sleep=1' );
+	$response = $http->get(
+		url: 'http://127.0.0.1:7878/?sleep=1',
+		options: [ 'timeout' => 1 ]
+	);
 
 	expect( $response['error'] )->toBe( true );
 	expect( $response['response_code'] )->toBe( 0 );