+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## ????
 - Default headers: connection, accept, user-agent
+- Can provide options on each verb method
 
 ## 0.0.1 : 2023-12-02
 - Basic working version